Updated entity constructor, renames EscapeRoom and builder with new constructors
