+++ S303-EscapeRoom/src/main/java/org/example/Modules/Entities/Entity.java
@@ -6,7 +6,11 @@ public class Entity {
     private int deleted;
     private int id;
 
-    public Entity (int deleted) {
+    public Entity () {
+        this.deleted = 0;
+    }
+    public Entity(int id, int deleted) {
+        this.id = id;
         this.deleted = deleted;
     }
     public ArrayList<String> getValues(){
